initial commit
I had already been working on this for about a week before my friend convinced me to make a github repo.

I am very much invested in the future of this project, and I wish to preserve it as best as possible.

So yeah. There's already a bunch of stuff here.