fixed gitignore, added some stuff to audiohandler
No longer including libsndfile dlls or libs in the remote project.

Debugging environment apparently requires dlls to be in the solution directory when using more than one environment (at the very least, that was the only fix I tried that worked).

beginning to implement audio stuff. it doesn't work yet, but we're getting there.